refactor(server): moved socket logic to a separate controller
